fixed after changes(all working) all refactored
