+++ tests/constructor-bottom/constructor-bottom.variables.js
@@ -1,34 +1,8 @@
 const { gql } = require('@apollo/client');
 const { setupApp } = require('../helper-functions');
-let operations;
-
-const wrongID = '6009dcd5f9855555907ebf5e';
-
-const color = {
-  name: [
-    { lang: 'ua', value: 'Світsadas' },
-    { lang: 'en', value: 'blackdasdas' },
-  ],
-  colorHex: '#f47ac6',
-  simpleName: [
-    { lang: 'ua', value: 'Чорний' },
-    { lang: 'en', value: 'black' },
-  ],
-};
-
-const newMaterial = colorID => ({
-  name: [
-    { lang: 'ua', value: 'тканина кордура' },
-    { lang: 'en', value: 'Cordura fabric' },
-  ],
-  description: [
-    { lang: 'ua', value: 'описання' },
-    { lang: 'en', value: 'some description' },
-  ],
-  purpose: 'INNER',
-  colors: colorID,
-  available: true,
-});
+const { deleteColor } = require('../color/color.helper');
+const { deleteConstructorBottom } = require('./constructor-bottom.helper');
+const { deleteMaterial } = require('../materials/material.helper');
 
 const newConstructorBottom = (colorId, materialId) => ({
   name: [
@@ -63,58 +37,9 @@ const getConstructorData = construrtor => ({
 
 const deleteAll = async (colorID, materialID, constructorBottomID) => {
   const operations = await setupApp();
-
-  const deleteConstructorBottom = await operations.mutate({
-    mutation: gql`
-      mutation($id: ID!) {
-        deleteConstructorBottom(id: $id) {
-          ... on ConstructorBottom {
-            _id
-          }
-          ... on Error {
-            statusCode
-            message
-          }
-        }
-      }
-    `,
-    variables: { id: constructorBottomID },
-  });
-  const deleteMaterial = await operations.mutate({
-    mutation: gql`
-      mutation($id: ID!) {
-        deleteMaterial(id: $id) {
-          ... on Material {
-            _id
-          }
-          ... on Error {
-            statusCode
-            message
-          }
-        }
-      }
-    `,
-    variables: { id: materialID },
-  });
-
-  const deleteColor = await operations.mutate({
-    mutation: gql`
-      mutation($id: ID!) {
-        deleteColor(id: $id) {
-          __typename
-          ... on Color {
-            _id
-          }
-          ... on Error {
-            statusCode
-            message
-          }
-        }
-      }
-    `,
-    variables: { id: colorID },
-  });
-
+  await deleteConstructorBottom(constructorBottomID, operations);
+  await deleteMaterial(materialID, operations);
+  await deleteColor(colorID, operations);
   return { deleteColor, deleteMaterial, deleteConstructorBottom };
 };
 
@@ -131,42 +56,9 @@ const getConstructorDataForUpt = (materialId, colorId) => ({
   default: true,
 });
 
-const createConstructorBottomQuery = async addConstructor => {
-  const operations = await setupApp();
-  const createConstructorBottom = await operations.mutate({
-    mutation: gql`
-      mutation($constructorElement: ConstructorBottomInput!) {
-        addConstructorBottom(constructorElement: $constructorElement) {
-          ... on ConstructorBottom {
-            _id
-            name {
-              lang
-              value
-            }
-            material {
-              _id
-            }
-            image
-            color {
-              _id
-            }
-            available
-            default
-          }
-        }
-      }
-    `,
-    variables: { constructorElement: addConstructor },
-  });
-  return createConstructorBottom.data.addConstructorBottom._id;
-};
 module.exports = {
-  color,
-  wrongID,
-  newMaterial,
   newConstructorBottom,
   getConstructorData,
   deleteAll,
   getConstructorDataForUpt,
-  createConstructorBottomQuery,
 };
